% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_manatee.R
\name{geom_manatee}
\alias{geom_manatee}
\title{geom_manatee}
\usage{
geom_manatee(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  inherit.aes = TRUE,
  na.rm = FALSE,
  by = "width",
  nudge_x = 0,
  ...
)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{inherit.aes}{logical, whether inherit aes from ggplot()}

\item{na.rm}{logical, whether remove NA values}

\item{by}{one of 'width' or 'height'}

\item{nudge_x}{horizontal adjustment to nudge manatees}

\item{...}{additional parameters}
}
\value{
geom layer
}
\description{
geom layer adding manatees
}
\examples{
library("ggplot2")
ggplot(mtcars) +
geom_manatee(aes(mpg, wt), manatee = "manatee", size = 5)

set.seed(1)
df <- data.frame(x = rnorm(10),
                 y = rnorm(10),
                 image = sample(c("manatee",
                                  "pusheen",
                                  "colonel",
                                  "venus",
                                  "toast"),
                                 size = 10, replace = TRUE))
 ggplot(df) +
geom_manatee(aes(x, y, manatee = image), size = 5)

}
